move view model factories to sdk; rename ChannelViewModelsFactory to ChatViewModelFactory
