+++ stream-chat-android-sample/src/main/kotlin/io/getstream/chat/sample/feature/channels/ChannelsFragment.kt
@@ -10,12 +10,14 @@ import androidx.navigation.fragment.findNavController
 import com.getstream.sdk.chat.viewmodel.channels.ChannelsViewModel
 import com.getstream.sdk.chat.viewmodel.channels.ChannelsViewModelImpl
 import com.getstream.sdk.chat.viewmodel.channels.bindView
+import com.getstream.sdk.chat.viewmodel.factory.ChannelsViewModelFactory
 import io.getstream.chat.sample.R
 import io.getstream.chat.sample.common.navigateSafely
 import kotlinx.android.synthetic.main.fragment_channels.*
 
 class ChannelsFragment : Fragment(R.layout.fragment_channels) {
-    private val viewModel: ChannelsViewModelImpl by viewModels()
+
+    private val viewModel: ChannelsViewModelImpl by viewModels { ChannelsViewModelFactory() }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         viewModel.bindView(channelsListView, viewLifecycleOwner)
