+++ stream-chat-android/src/main/java/com/getstream/sdk/chat/viewmodel/factory/ChatViewModelFactory.kt
@@ -1,4 +1,4 @@
-package io.getstream.chat.sample.feature.channel
+package com.getstream.sdk.chat.viewmodel.factory
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
@@ -7,7 +7,8 @@ import com.getstream.sdk.chat.viewmodel.MessageInputViewModel
 import com.getstream.sdk.chat.viewmodel.messages.MessageListViewModel
 
 @Suppress("UNCHECKED_CAST")
-class ChannelViewModelsFactory(private val cid: String) : ViewModelProvider.NewInstanceFactory() {
+public class ChatViewModelFactory(private val cid: String) :
+    ViewModelProvider.NewInstanceFactory() {
     override fun <T : ViewModel?> create(modelClass: Class<T>): T = when (modelClass) {
         MessageListViewModel::class.java -> MessageListViewModel(cid) as T
         ChannelHeaderViewModel::class.java -> ChannelHeaderViewModel(cid) as T
